(PC-25763) refactor(TabLayout): Move info tab to new file
