+++ src/features/venue/components/TabLayout/TabLayout.tsx
@@ -1,25 +1,19 @@
 import React, { FunctionComponent, useRef, useState } from 'react'
-import { Platform, View } from 'react-native'
+import { View } from 'react-native'
 import styled from 'styled-components/native'
 
-import { TouchableTab } from 'features/venue/components/TabLayout/TouchableTab'
 import { useTabArrowNavigation } from 'features/venue/components/TabLayout/useTabArrowNavigation'
 import { Tab } from 'features/venue/types'
 import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
-import { useHandleHover } from 'libs/hooks/useHandleHover'
-import { getSpacing, Spacer, Typo } from 'ui/theme'
+import { getSpacing, Spacer } from 'ui/theme'
+
+import { InfoTab } from './InfoTab'
 
 const tabs = Object.values(Tab)
 
 type Props = {
   tabPanels: Record<Tab, JSX.Element>
   onTabChange?: Record<Tab, () => void>
-}
-
-type InfoTabProps = {
-  tab: Tab
-  selectedTab: Tab
-  onPress: () => void
 }
 
 export const TabLayout: FunctionComponent<Props> = ({ tabPanels, onTabChange }) => {
@@ -48,25 +42,6 @@ export const TabLayout: FunctionComponent<Props> = ({ tabPanels, onTabChange }) 
   )
 }
 
-const InfoTab = ({ tab, selectedTab, onPress }: InfoTabProps) => {
-  const isSelected = selectedTab === tab
-  const { isHover, ...webHoverProps } = useHandleHover()
-  const hoverProps = Platform.OS === 'web' ? webHoverProps : {}
-
-  return (
-    <StyledTouchableTab id={tab} onPress={onPress} selected={isSelected} {...hoverProps}>
-      <Spacer.Column numberOfSpaces={6} />
-      <TabTitleContainer>
-        <TabTitle isHover={isHover} isSelected={isSelected}>
-          {tab}
-        </TabTitle>
-      </TabTitleContainer>
-      <Spacer.Column numberOfSpaces={2} />
-      <BarOfSelectedTab isSelected={isSelected} />
-    </StyledTouchableTab>
-  )
-}
-
 const Container = styled.View({
   width: '100%',
 })
@@ -75,32 +50,10 @@ const TabContainer = styled.View({
   flexDirection: 'row',
 })
 
-const StyledTouchableTab = styled(TouchableTab)({
-  flex: 1,
-  maxWidth: getSpacing(45),
-})
-
 const GreyBar = styled.View(({ theme }) => ({
   position: 'absolute',
   bottom: 0,
   height: getSpacing(1),
   width: '100%',
   backgroundColor: theme.colors.greyLight,
-}))
-
-const TabTitleContainer = styled.View({ flexGrow: 1, justifyContent: 'center' })
-
-const TabTitle = styled(Typo.ButtonText)<{ isSelected: boolean; isHover: boolean }>(
-  ({ isSelected, isHover, theme }) => ({
-    textAlign: 'center',
-    color: isSelected || isHover ? theme.colors.primary : theme.colors.greyDark,
-  })
-)
-
-const BarOfSelectedTab = styled.View<{ isSelected: boolean }>(({ theme, isSelected }) => ({
-  bottom: 0,
-  height: getSpacing(1),
-  width: '100%',
-  backgroundColor: isSelected ? theme.colors.primary : 'transparent',
-  borderRadius: getSpacing(1),
 }))
